Shift squares' center up and to the left
Previously, the edges of the squares were flush against the bounds of
the viewing window. I think that consequently, for some circumstances
ggplot2 was truncating the outer square if it was arbitrarily too close
to the edge. By nudging the squares slightly off those edges, they
should retain almost the same position but no longer be truncated.

Fixes #45