Change granovagg.1w squares color when F is significant
Create granovagg.1w-helpers file to store helper functions for the
one-way function.

Add tests for IsFSignificant(). All tests pass.

Previously, the squares color was determined by whether or not the F
statistic was greater than 1. I think it's more useful to have a visual
indicator of whether or not one's data exceeds the critical F-statistic
for alpha = 0.05.

Closes #119