Remove extra version information string
We used to uniquely version each build every commit on dev. It was a pre-commit hook. But I think the extra version information is causing an error with Roxygen. And, because we can just use continuous integration builds for build versions, we no longer need to encode a unique id into each version number.